Replace Snyk container scanning by Trivy
Trivy [1] is the currently recommended tool for container scanning since
Snyk is being phased out. We pin the version of the Trivy Action [2] to
its most recent release as recommended for third party Actions [3]. This
is recommended to safeguard against supply chain attacks.

We mimic both, Snyk's scanning of configuration files for infrastructure
as code (radixconfig and Docker) and scanning of Docker images. The
Docker image generated on the `radix` branch is also scanned.

Reporting of the results depends on which branch is scanned. If the scan
is run on the `main` or `radix` branch, the reports are uploaded to
GitHub and presented in the "Security" tab of the repository. In all
other cases, the results are reported within the Actions output and the
action fails if an error is found.

The container images are only build and scanned for the `main` and
`radix` branch to save resources and to avoid unnecessary noise inside
pull requests.

[1]: https://github.com/aquasecurity/trivy
[2]: https://github.com/aquasecurity/trivy-action
[3]: https://docs.github.com/en/actions/reference/security/secure-use#using-third-party-actions