Replace Snyk container scanning by Trivy
Trivy [1] is the currently recommended tool for container scanning since
Snyk is being phased out. We pin the version of the Trivy Action [2] to
its most recent release as recommended for third party Actions [3]. This
is recommended to safeguard against supply chain attacks.

We only scan the container image that is used deploying the oneseismic
API. Containers only used for testing and thus not permanently used are
currently not scanned to save resoruces.

[1]: https://github.com/aquasecurity/trivy
[2]: https://github.com/aquasecurity/trivy-action
[3]: https://docs.github.com/en/actions/reference/security/secure-use#using-third-party-actions